⚡ Bolt: Optimize MessageList re-renders
Extracted MessageItem component and wrapped it in React.memo to prevent re-rendering of all messages when a new token is streamed. This significantly reduces CPU usage during chat streaming, especially with markdown and syntax highlighting.
