+++ web/src/components/MessageItem.tsx
@@ -1,0 +1,111 @@
+import React from 'react';
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+import remarkMath from 'remark-math';
+import rehypeKatex from 'rehype-katex';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { vscDarkPlus } from 'react-syntax-highlighter/dist/esm/styles/prism';
+import 'katex/dist/katex.min.css';
+import { Message } from './ChatInterface.types';
+
+interface MessageItemProps {
+  message: Message;
+  index: number;
+  isFolded: boolean;
+  onToggleFold: (index: number, currentState: boolean) => void;
+  setRef: (el: HTMLDivElement | null, index: number) => void;
+}
+
+export const MessageItem = React.memo(({ message, index, isFolded, onToggleFold, setRef }: MessageItemProps) => {
+  const messageClass =
+    message.role === 'user'
+      ? 'user-message'
+      : message.type === 'tool_calls' || message.type === 'tool_output'
+        ? 'tool-message'
+        : message.type === 'reasoning_summary'
+          ? 'reasoning-message'
+          : 'assistant-message';
+
+  const isFoldable = message.type === 'tool_calls' || message.type === 'tool_output' || message.type === 'reasoning_summary';
+
+  return (
+    <div
+      className={`message ${messageClass} ${isFoldable && isFolded ? 'folded-clickable' : ''}`}
+      ref={(el) => setRef(el, index)}
+      onClick={() => {
+        if (isFoldable && isFolded) {
+          onToggleFold(index, isFolded);
+        }
+      }}
+    >
+      {/* For foldable messages (tool_calls, reasoning_summary), show a toggle */}
+      {isFoldable && (
+        <button
+          type="button"
+          className="tool-summary"
+          onClick={(event) => {
+            event.stopPropagation();
+            onToggleFold(index, isFolded);
+          }}
+        >
+          <span className="fold-arrow" aria-hidden="true">
+            {isFolded ? '⇨' : '⇩'}
+          </span>
+          <strong>
+            {message.type === 'tool_calls'
+              ? 'Tool Call'
+              : message.type === 'tool_output'
+                ? 'Tool Output'
+                : message.type === 'reasoning_summary'
+                  ? 'Reasoning Summary'
+                  : 'Message'}
+          </strong>
+        </button>
+      )}
+
+      {/* Show content unless folded */}
+      {!isFolded && (
+        <>
+          {message.images && message.images.length > 0 && (
+            <div className="message-images-container">
+              {message.images.map((imgSrc, imgIndex) => (
+                <div key={imgIndex} className="message-image-container">
+                  <img src={imgSrc} alt="Message attachment" className="message-image" />
+                </div>
+              ))}
+            </div>
+          )}
+          <div className="message-text markdown-content">
+            <ReactMarkdown
+              remarkPlugins={[remarkGfm, remarkMath]}
+              rehypePlugins={[rehypeKatex]}
+              components={{
+                p: ({ ...props }) => <p className="tight-paragraph" {...props} />,
+                li: ({ ...props }) => <li className="tight-list-item" {...props} />,
+                code: (props: { inline?: boolean; className?: string; children?: React.ReactNode }) => {
+                  const { inline, className, children, ...rest } = props;
+                  const cls = className || '';
+                  const langToken = cls.split(' ').find((c) => c.startsWith('language-'));
+                  const lang = !inline && langToken ? langToken.replace('language-', '') : undefined;
+                  return !inline && lang ? (
+                    <SyntaxHighlighter style={vscDarkPlus as Record<string, React.CSSProperties>} language={lang} PreTag="div" {...rest}>
+                      {String(children).replace(/\n$/, '')}
+                    </SyntaxHighlighter>
+                  ) : (
+                    <code className={className} {...rest}>
+                      {children}
+                    </code>
+                  );
+                },
+              }}
+            >
+              {message.content}
+            </ReactMarkdown>
+          </div>
+        </>
+      )}
+    </div>
+  );
+});
+
+MessageItem.displayName = 'MessageItem';
