+++ web/src/components/MessageList.tsx
@@ -1,12 +1,6 @@
-import React, { useState, useEffect, useRef } from 'react';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import remarkMath from 'remark-math';
-import rehypeKatex from 'rehype-katex';
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { vscDarkPlus } from 'react-syntax-highlighter/dist/esm/styles/prism';
-import 'katex/dist/katex.min.css';
+import React, { useState, useEffect, useRef, useCallback } from 'react';
 import { Message } from './ChatInterface.types';
+import { MessageItem } from './MessageItem';
 
 interface MessageListProps {
   messages: Message[];
@@ -17,19 +11,24 @@ export const MessageList: React.FC<MessageListProps> = ({ messages, isLoading })
   const [foldedMessages, setFoldedMessages] = useState<Record<string, boolean>>({});
   const messageRefs = useRef<Record<number, HTMLDivElement | null>>({});
 
-  const toggleFolded = (index: number, currentState: boolean) => {
+  const toggleFolded = useCallback((index: number, currentState: boolean) => {
     setFoldedMessages((prev) => ({
       ...prev,
       [index]: !currentState,
     }));
-  };
+  }, []);
+
+  const setRef = useCallback((el: HTMLDivElement | null, index: number) => {
+    messageRefs.current[index] = el;
+  }, []);
 
   // Auto-scroll logic for streaming long messages
   useEffect(() => {
     const lastMessageIndex = messages.length - 1;
     if (lastMessageIndex >= 0) {
       const lastMessage = messages[lastMessageIndex];
-      const isLongMessage = lastMessage.type === 'tool_calls' || lastMessage.type === 'tool_output' || lastMessage.type === 'reasoning_summary';
+      const isLongMessage =
+        lastMessage.type === 'tool_calls' || lastMessage.type === 'tool_output' || lastMessage.type === 'reasoning_summary';
 
       const isFolded = foldedMessages[lastMessageIndex] ?? lastMessage.folded ?? false;
 
@@ -54,97 +53,16 @@ export const MessageList: React.FC<MessageListProps> = ({ messages, isLoading })
     <div className="messages-list">
       {messages.map((message, index) => {
         const isFolded = foldedMessages[index] ?? message.folded ?? false;
-        const messageClass =
-          message.role === 'user'
-            ? 'user-message'
-            : message.type === 'tool_calls' || message.type === 'tool_output'
-              ? 'tool-message'
-              : message.type === 'reasoning_summary'
-                ? 'reasoning-message'
-                : 'assistant-message';
-
-        const isFoldable = message.type === 'tool_calls' || message.type === 'tool_output' || message.type === 'reasoning_summary';
 
         return (
-          <div
-            key={index}
-            className={`message ${messageClass} ${isFoldable && isFolded ? 'folded-clickable' : ''}`}
-            ref={(el) => {
-              messageRefs.current[index] = el;
-            }}
-            onClick={() => {
-              if (isFoldable && isFolded) {
-                toggleFolded(index, isFolded);
-              }
-            }}
-          >
-            {/* For foldable messages (tool_calls, reasoning_summary), show a toggle */}
-            {isFoldable && (
-              <button
-                type="button"
-                className="tool-summary"
-                onClick={(event) => {
-                  event.stopPropagation();
-                  toggleFolded(index, isFolded);
-                }}
-              >
-                <span className="fold-arrow" aria-hidden="true">
-                  {isFolded ? '⇨' : '⇩'}
-                </span>
-                <strong>
-                  {message.type === 'tool_calls'
-                    ? 'Tool Call'
-                    : message.type === 'tool_output'
-                      ? 'Tool Output'
-                      : message.type === 'reasoning_summary'
-                        ? 'Reasoning Summary'
-                        : 'Message'}
-                </strong>
-              </button>
-            )}
-
-            {/* Show content unless folded */}
-            {!isFolded && (
-              <>
-                {message.images && message.images.length > 0 && (
-                  <div className="message-images-container">
-                    {message.images.map((imgSrc, imgIndex) => (
-                      <div key={imgIndex} className="message-image-container">
-                        <img src={imgSrc} alt="Message attachment" className="message-image" />
-                      </div>
-                    ))}
-                  </div>
-                )}
-                <div className="message-text markdown-content">
-                  <ReactMarkdown
-                    remarkPlugins={[remarkGfm, remarkMath]}
-                    rehypePlugins={[rehypeKatex]}
-                    components={{
-                      p: ({ ...props }) => <p className="tight-paragraph" {...props} />,
-                      li: ({ ...props }) => <li className="tight-list-item" {...props} />,
-                      code: (props: { inline?: boolean; className?: string; children?: React.ReactNode }) => {
-                        const { inline, className, children, ...rest } = props;
-                        const cls = className || '';
-                        const langToken = cls.split(' ').find((c) => c.startsWith('language-'));
-                        const lang = !inline && langToken ? langToken.replace('language-', '') : undefined;
-                        return !inline && lang ? (
-                          <SyntaxHighlighter style={vscDarkPlus as Record<string, React.CSSProperties>} language={lang} PreTag="div" {...rest}>
-                            {String(children).replace(/\n$/, '')}
-                          </SyntaxHighlighter>
-                        ) : (
-                          <code className={className} {...rest}>
-                            {children}
-                          </code>
-                        );
-                      },
-                    }}
-                  >
-                    {message.content}
-                  </ReactMarkdown>
-                </div>
-              </>
-            )}
-          </div>
+          <MessageItem
+            key={message.messageId || index}
+            message={message}
+            index={index}
+            isFolded={isFolded}
+            onToggleFold={toggleFolded}
+            setRef={setRef}
+          />
         );
       })}
       {isLoading && (
